+++ frontend/src/themes/index.ts
@@ -3,7 +3,9 @@ import { everforestTheme } from './everforestTheme'
 import { nordTheme } from './nordTheme'
 import { catppuccinTheme } from './catppuccinTheme'
 import { scienceTheme } from './scienceTheme'
+import { gruvboxTheme } from './gruvboxTheme'
+import { solarizedTheme } from './solarizedTheme'
 
-export { defaultTheme, everforestTheme, nordTheme, catppuccinTheme, scienceTheme }
+export { defaultTheme, everforestTheme, nordTheme, catppuccinTheme, scienceTheme, gruvboxTheme, solarizedTheme }
 
-export const themes = [defaultTheme, everforestTheme, nordTheme, catppuccinTheme, scienceTheme]
+export const themes = [defaultTheme, everforestTheme, nordTheme, catppuccinTheme, scienceTheme, gruvboxTheme, solarizedTheme]